added SystemUtils; moved system methods from FileUtils to SystemUtils; Added logic for window size calculation;
