+++ src/main/java/tz/deadparrot/utils/FileUtils.java
@@ -14,6 +14,7 @@ import java.time.LocalDateTime;
 
 @Slf4j
 public class FileUtils {
+
     public static void verifyAndCreateOutputFolder(String dirPath) {
         try {
             Path path = Paths.get(dirPath);
@@ -25,21 +26,6 @@ public class FileUtils {
             }
         } catch (Exception e) {
             log.error(Constants.ERROR_CREATING_FOLDER + "{}", e.getMessage());
-        }
-    }
-
-    public static void detectOS(Boolean showLogs) {
-        String osName = System.getProperty("os.name").toLowerCase();
-        if (osName.contains(Constants.LINUX_OS)) {
-            if (showLogs) log.info(Constants.RUNNING_LINUX);
-            Constants.IS_LINUX = true;
-            Settings.OUTPUT_DESKTOP_FOLDER_PATH = Constants.LINUX_DESKTOP_PATH;
-        } else if (osName.contains(Constants.WINDOWS_OS)) {
-            if (showLogs) log.info(Constants.RUNNING_WINDOWS);
-            Constants.IS_WINDOWS = true;
-            Settings.OUTPUT_DESKTOP_FOLDER_PATH = Constants.WINDOWS_DESKTOP_PATH;
-        } else {
-            Settings.SAVE_RECORDINGS_TO_DESKTOP = false;
         }
     }
